+++ packages/graphql-yoga/__tests__/subscriptions.spec.ts
@@ -29,7 +29,6 @@ function eventStream<TType = unknown>(source: ReadableStream<Uint8Array>) {
 
 describe('Subscription', () => {
   test('eventStream', async () => {
-    // eslint-disable-next-line @typescript-eslint/require-await
     const source = (async function* foo() {
       yield { hi: 'hi' }
       yield { hi: 'hello' }
@@ -86,4 +85,178 @@ describe('Subscription', () => {
       }
     }
   })
+
+  test('should issue pings while connected', async () => {
+    const d = createDeferred()
+
+    const schema = createSchema({
+      typeDefs: /* GraphQL */ `
+        type Subscription {
+          hi: String!
+        }
+        type Query {
+          hi: String!
+        }
+      `,
+      resolvers: {
+        Subscription: {
+          hi: {
+            async *subscribe() {
+              await d.promise
+              yield { hi: 'hi' }
+            },
+          },
+        },
+      },
+    })
+
+    const yoga = createYoga({ schema })
+
+    const response = await yoga.fetch('http://yoga/graphql', {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+        accept: 'text/event-stream',
+      },
+      body: JSON.stringify({
+        query: /* GraphQL */ `
+          subscription {
+            hi
+          }
+        `,
+      }),
+    })
+
+    const iterator = response.body![Symbol.asyncIterator]()
+
+    const results = []
+    let value: Uint8Array
+
+    while (({ value } = await iterator.next())) {
+      if (value === undefined) {
+        break
+      }
+      results.push(Buffer.from(value).toString('utf-8'))
+      if (results.length === 3) {
+        d.resolve()
+      }
+    }
+
+    expect(results).toMatchInlineSnapshot(`
+        [
+          ":
+
+        ",
+          ":
+
+        ",
+          ":
+
+        ",
+          "data: {"data":{"hi":"hi"}}
+
+        ",
+        ]
+      `)
+  })
+
+  test('should issue pings event if event source never publishes anything', async () => {
+    const d = createDeferred()
+    const source: AsyncIterableIterator<unknown> = {
+      next: () => d.promise.then(() => ({ done: true, value: undefined })),
+      return: () => {
+        d.resolve()
+        return Promise.resolve({ done: true, value: undefined })
+      },
+      throw: () => {
+        throw new Error('Method not implemented. (throw)')
+      },
+      [Symbol.asyncIterator]() {
+        return this
+      },
+    }
+
+    const schema = createSchema({
+      typeDefs: /* GraphQL */ `
+        type Subscription {
+          hi: String!
+        }
+        type Query {
+          hi: String!
+        }
+      `,
+      resolvers: {
+        Subscription: {
+          hi: {
+            subscribe: () => source,
+          },
+        },
+      },
+    })
+
+    const yoga = createYoga({ schema })
+
+    const response = await yoga.fetch('http://yoga/graphql', {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+        accept: 'text/event-stream',
+      },
+      body: JSON.stringify({
+        query: /* GraphQL */ `
+          subscription {
+            hi
+          }
+        `,
+      }),
+    })
+
+    const iterator = response.body![Symbol.asyncIterator]()
+
+    const results = []
+    let value: Uint8Array
+
+    while (({ value } = await iterator.next())) {
+      if (value === undefined) {
+        break
+      }
+      results.push(Buffer.from(value).toString('utf-8'))
+      if (results.length === 4) {
+        await iterator.return!()
+      }
+    }
+
+    await d.promise
+    expect(results).toMatchInlineSnapshot(`
+      [
+        ":
+
+      ",
+        ":
+
+      ",
+        ":
+
+      ",
+        ":
+
+      ",
+      ]
+    `)
+  })
 })
+
+type Deferred<T = void> = {
+  resolve: (value: T) => void
+  reject: (value: unknown) => void
+  promise: Promise<T>
+}
+
+function createDeferred<T = void>(): Deferred<T> {
+  const d = {} as Deferred<T>
+  d.promise = new Promise<T>((resolve, reject) => {
+    d.resolve = resolve
+    d.reject = reject
+  })
+  return d
+}