+++ packages/graphql-yoga/src/plugins/resultProcessor/push.ts
@@ -4,13 +4,18 @@ import { isAsyncIterable } from '@envelop/core'
 import { getResponseInitByRespectingErrors } from '../../error.js'
 import { FetchAPI, MaybeArray } from '../../types.js'
 import { ResultProcessorInput } from '../types.js'
-import { jsonStringifyResult } from './stringify.js'
+import { jsonStringifyResultWithoutInternals } from './stringify.js'
 
 export function processPushResult(
   result: ResultProcessorInput,
   fetchAPI: FetchAPI,
 ): Response {
-  const timeoutInSeconds = 12
+  let pingIntervalMs = 12_000
+
+  // for testing the pings, reduce the timeout
+  if (globalThis.process?.env?.NODE_ENV === 'test') {
+    pingIntervalMs = 5
+  }
 
   const headersInit = {
     'Content-Type': 'text/event-stream',
@@ -33,8 +38,9 @@ export function processPushResult(
           clearInterval(pingInterval)
           return
         }
+
         controller.enqueue(textEncoder.encode(':\n\n'))
-      }, timeoutInSeconds * 1000) as unknown as number
+      }, pingIntervalMs) as unknown as number
 
       if (isAsyncIterable(result)) {
         iterator = result[Symbol.asyncIterator]()
@@ -53,8 +59,9 @@ export function processPushResult(
     },
     async pull(controller) {
       const { done, value } = await iterator.next()
+
       if (value != null) {
-        const chunk = jsonStringifyResult(value)
+        const chunk = jsonStringifyResultWithoutInternals(value)
         controller.enqueue(textEncoder.encode(`data: ${chunk}\n\n`))
       }
       if (done) {