+++ packages/graphql-yoga/src/plugins/useParserAndValidationCache.ts
@@ -1,5 +1,11 @@
-import { memoize2of4 } from '@graphql-tools/utils'
-import type { DocumentNode, parse, validate } from 'graphql'
+import type { AfterValidateHook } from '@envelop/core'
+import type {
+  DocumentNode,
+  GraphQLError,
+  GraphQLSchema,
+  validate,
+  ValidationRule,
+} from 'graphql'
 import { createLRUCache } from '../utils/create-lru-cache.js'
 import type { Plugin } from './types.js'
 
@@ -20,54 +26,63 @@ export function useParserAndValidationCache({
   validationCache = true,
 }: // eslint-disable-next-line @typescript-eslint/ban-types
 ParserAndValidationCacheOptions): Plugin<{}> {
-  const memoizedValidateByRules =
-    typeof validationCache === 'boolean'
-      ? createLRUCache<typeof validate>()
-      : validationCache
+  const validationCacheByRules =
+    createLRUCache<
+      WeakMap<GraphQLSchema, WeakMap<DocumentNode, GraphQLError[]>>
+    >()
   return {
-    onParse({
-      parseFn,
-      setParseFn,
-    }: {
-      parseFn: typeof parse
-      setParseFn: (fn: typeof parse) => void
-    }) {
-      setParseFn(function memoizedParse(source) {
-        const strDocument = typeof source === 'string' ? source : source.body
-        let document = documentCache.get(strDocument)
-        if (!document) {
-          const parserError = errorCache.get(strDocument)
-          if (parserError) {
-            throw parserError
+    onParse({ params, setParsedDocument }) {
+      const strDocument = params.source.toString()
+      const document = documentCache.get(strDocument)
+      if (document) {
+        setParsedDocument(document)
+        return
+      }
+      const parserError = errorCache.get(strDocument)
+      if (parserError) {
+        throw parserError
+      }
+      return ({ result }) => {
+        if (result != null) {
+          if (result instanceof Error) {
+            errorCache.set(strDocument, result)
+          } else {
+            documentCache.set(strDocument, result)
           }
-          try {
-            document = parseFn(source)
-          } catch (e) {
-            errorCache.set(strDocument, e)
-            throw e
-          }
-          documentCache.set(strDocument, document)
         }
-        return document
-      })
+      }
     },
     onValidate({
-      validateFn,
-      setValidationFn,
-    }: {
-      validateFn: typeof validate
-      setValidationFn: (fn: typeof validate) => void
-    }) {
+      params: { schema, documentAST, rules },
+      setResult,
+      // eslint-disable-next-line @typescript-eslint/ban-types
+    }): void | AfterValidateHook<{}> {
       if (validationCache !== false) {
-        setValidationFn(function memoizedValidateFn(schema, document, rules) {
-          const rulesKey = rules?.map((rule) => rule.name).join(',') || ''
-          let memoizedValidateFnForRules = memoizedValidateByRules.get(rulesKey)
-          if (!memoizedValidateFnForRules) {
-            memoizedValidateFnForRules = memoize2of4(validateFn)
-            memoizedValidateByRules.set(rulesKey, memoizedValidateFnForRules)
+        const rulesKey =
+          rules?.map((rule: ValidationRule) => rule.name).join(',') || ''
+        let validationCacheBySchema = validationCacheByRules.get(rulesKey)
+        if (!validationCacheBySchema) {
+          validationCacheBySchema = new WeakMap()
+          validationCacheByRules.set(rulesKey, validationCacheBySchema)
+        }
+        let validationCacheByDocument = validationCacheBySchema.get(schema)
+        if (!validationCacheByDocument) {
+          validationCacheByDocument = new WeakMap()
+          validationCacheBySchema.set(schema, validationCacheByDocument)
+        }
+        const cachedResult = validationCacheByDocument.get(documentAST)
+        if (cachedResult) {
+          setResult(cachedResult)
+          return
+        }
+        return ({ result }) => {
+          if (result != null) {
+            validationCacheByDocument?.set(
+              documentAST,
+              result as GraphQLError[],
+            )
           }
-          return memoizedValidateFnForRules(schema, document, rules)
-        })
+        }
       }
     },
   }